requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pgvector
pillow
numpy
torch  
torchvision
python-multipart